@@ -1,9 +1,13 @@
 import './App.css';
+import { NewHomes, NavBar, MarketingFooter } from './ui-components'
 
 function App() {
   return (
-    <div className="App">
-    </div>
+<div className="App">
+  <NavBar />
+  <NewHomes />
+  <MarketingFooter />
+</div>
   );
 }
 
